SFO-102: Fix conflict with Josue
